getAllDetections, getAllPeople routes
writing jest tests to pass through session validation was proving difficult so i put it on the back burner in favor of the new routes. All three of the routes today have been manually tested in with postman and their effects in the mongo database have been monitored. unit tests pending.